Dev back

+++ api_server/routers/admins.py
@@ -15,8 +15,17 @@ def get_admins_service() -> AdminServiceModule.AdminService:
         
 @router.get("/protected")
 async def read_protected_route(current_admin: admins_model.TokenData = Depends(get_current_admin)):
-    return {"username": current_admin.username,"role":current_admin.role,"account":current_admin.account, "message": "Welcome to a protected route!"}
-
+    try:
+        return {"username": current_admin.username,"role":current_admin.role,"account":current_admin.account, "message": "Welcome to a protected route!"}
+    except HTTPException as he:
+        raise he
+    except Exception as e:
+        logger.error(f"Unexpected error: {str(e)}")
+        # 開法者模式使用
+        # return {"error": str(e)}
+        # 這裡捕獲了任何其他的異常
+        raise HTTPException(detail="Server error",status_code=500)
+    
 @router.get("/{admin_id}")
 async def read_admin(admin_id: str,admins_service: AdminServiceModule.AdminService = Depends(get_admins_service),current_admin: admins_model.TokenData = Depends(get_current_admin)):
     try:
@@ -63,11 +72,15 @@ async def create_admin(admin: admins_model.AdminCreate,admins_service: AdminServ
 @router.patch("/{admin_id}")
 async def update_admin(admin_id: str,admin: admins_model.AdminUpdate,admins_service: AdminServiceModule.AdminService = Depends(get_admins_service),current_admin: admins_model.TokenData = Depends(get_current_admin)):
     try:
+        if admin.role and current_admin.role!="admin":
+            raise HTTPException(status_code=403, detail="Permission denied")
         await admins_service.update_admin(admin_id,admin)
         return {"detail": "success"}
     except ValueError as ve:  
         logger.warning(f"Value error encountered: {str(ve)}")
         raise HTTPException(status_code=400, detail=str(ve))
+    except HTTPException as he:
+        raise he
     except Exception as e:
         logger.error(f"Unexpected error: {str(e)}")
         raise HTTPException(detail="Server error", status_code=500)
@@ -75,11 +88,15 @@ async def update_admin(admin_id: str,admin: admins_model.AdminUpdate,admins_serv
 @router.delete("/{admin_id}")
 async def delete_admin(admin_id: str,admins_service: AdminServiceModule.AdminService = Depends(get_admins_service),current_admin: admins_model.TokenData = Depends(get_current_admin)):
     try:
+        if current_admin.role!="admin":
+            raise HTTPException(status_code=403, detail="Permission denied")
         await admins_service.delete_admin(admin_id)
         return {"detail": "success"}
     except ValueError as ve:  
         logger.warning(f"Value error encountered: {str(ve)}")
         raise HTTPException(status_code=400, detail=str(ve))
+    except HTTPException as he:
+        raise he
     except Exception as e:
         logger.error(f"Unexpected error: {str(e)}")
         raise HTTPException(detail="Server error",status_code=500)
